* merge back changes from VST SDK 3.5:
		* fix some warnings with some compilers
		* add CAnimationSplashScreen class
		* animations can have an optional notification object for receiving a message when the animation has finished

+++ vstgui/lib/platform/win32/win32dragcontainer.cpp
@@ -181,7 +181,7 @@ void* WinDragContainer::next (int32_t& size, int32_t& type)
 				if (type == kUnicodeText)
 				{
 					UTF8StringHelper wideString ((const WCHAR*)data);
-					size = strlen (wideString.getUTF8String ());
+					size = (int32_t)strlen (wideString.getUTF8String ());
 					lastItem = malloc (size+1);
 					strcpy ((char*)lastItem, wideString.getUTF8String ());
 				}
